Change dataSize() to return a UINT64 to ensure products are promoted to UINT64 rather than overflowing an int. Addressed the phorum issue http://www.lemurproject.org/phorum/read.php?11,3655

+++ utility/include/indri/BulkTree.hpp
@@ -72,7 +72,7 @@ namespace indri
 
       void clear();
       char* data();
-      static int dataSize();
+      static UINT64 dataSize();
 
       // linked list
       void setID( UINT32 id );
